requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
PyMySQL==1.1.1
python-dotenv==1.0.1
waitress==3.0.0
